requirements.txt
flask>=3.0
lark-oapi>=1.3
openai>=1.0
pinecone>=5.0
python-dotenv>=1.0
